Fix improper array expansion
Positional arguments passed to filter where meant to be used as a single
argument to the inner shell, which splits them itself. This means that
it is not necessary to quote very simple predicates.

With "$@", `filter true false` would have expanded to the something like
the following:

```
xargs sh -c 'true' 'false >/dev/null && printf '\''%s\n'\'' "$1"' filter
```

where the command `true` is the predicate, `$1` contains the string
`"false >/dev/null && ..."`, `$2` contains `"filter"`, and `$3` contains
the actual input line.