Revert setting PATH in .profile
Setting PATH in ~/.profile has the unfortunate consequence that making a
change to the PATH requires in a desktop environment to logout and login
again to take effect. Setting the PATH in .commonrc allows to get the
updated value in all new shells.

On Alpine linux' Bash the default profile does not load any bashrc, so
we have to load it back in the user's .bash_profile. This might not work
as expected in all circumstances, cf. [Bash Initalization Behaviour].

[Bash Initalization Behaviour]: https://github.com/0cjs/sedoc/blob/master/lang/bash/init.md