Remove unnecessary sleep
Sleep is more precise than countdown, but that's not necessary for the
uses of zzz. Also the change makes the code prettier.

This required to definitively fix the weird issue with oninput, where
under conditions I don't fully understand it would not receive the
user's input. My current understanding is that it was actually strange
that it received the user's input as running a command in the background
should normally detach its stdin by default.

This revision also fixes an edge case where the screen could be turned
off after the script had terminated.