Update keybindings for thinkpad X1C8
The thinkpad X1C8 does not have media keys and has PGUP and PGDN on top of
resp. the left and right arrows, which makes them behave strangely with the
default mpv configuration (PGUP seeks the the next chapter, but on this layout
it should seek to the previous one).