Allow matching on ID, unless -e is specified
ID matching is triggered only if matching on names yield no result. This allows
to avoid collisions, and usually it would be sufficient to add an extra
character to the filter if one really wants to match an ID.

This commit also fixes matching on partial IDs when the '-e' flag is specified.