Use read loop instead of xargs
The "delimiter" of read is actually more of a terminator, and read won't run
the loop if the last line is not terminated by the "delimiter". So it must be
ensured that the last line is properly null-terminated.

The sed pattern to replace the newline between the end token of a certificate
and the start token of the following certificate is a bit intricate, as sed's
`N` command will end the script when the last line is already in the pattern
space. Furthermore, in GNU sed, `N` will print the content of the pattern space
(but the following commands, e.g. a substitution, still won't be applied), but
POSIX-compliant implementation do not output anything.

To work around those issues, the script first checks whether the last line is
being processed, and if so appends a null byte using the `s` command, then
explicitly prints the pattern space and terminates (like `N`, the `n` command
will end the script if the last line has already been reached, but unlike `N`
it will always print the pattern space).

To accomodate cases where it is not known whether the input ends with a newline
(it should always be the case here with openssl's output, but this snipped is
written to be reusable), the sed script `$a\` is used to append a newline if
there is none, but does nothing if there is already a newline. The command
seemingly "appends nothing" (as there is no text after the backslash), but
implicitly sed appends a newline to the lines it processes if not already
there. [1]

[1]: https://unix.stackexchange.com/a/31955