Fix nautilus not opening scripts anymore
After the last batch of system updates, nautilus would not execute scripts
anymore. Turns out it was xdg-open that couldn't handle the scripts anymore
because it seems that the mime type of scripts has changed.

This does not solve nautilus' new "Run as a Program" right-click menu entry
which still doesn't work with basically the same error as xdg-open previously:

    sh: line 1: exec: xterm: not found

Apparently nautilus uses GIO to handle this new action [1][2], and GIO has a
hardcoded list of terminals it is willing to use [3][4] defaulting to xterm
(hence the error message; I have none of the terminals in the hardoced list).

[1]: https://www.reddit.com/r/gnome/comments/2z1ai5/change_default_terminal_application/cpeu7sd?utm_source=share&utm_medium=web2x&context=3
[2]: https://www.reddit.com/r/swaywm/comments/fgk3w0/gio/?utm_source=share&utm_medium=web2x&context=3
[3]: https://gitlab.gnome.org/GNOME/glib/-/issues/338
[4]: https://gitlab.gnome.org/GNOME/glib/-/blob/master/gio/gdesktopappinfo.c#L2473