Fix for nautilus' "Run as Program"
When none of gnome-terminal, xterm, ... is installed, nautilus fails to run
scripts as programs from its new "Run as Program" menu.

This is caused by nautilus using GIO to execute those scripts, and GIO only
supports a very limited, hardcoded list of terminal emulators [1][2].

This shim redirects calls for gnome-terminal to terminator. It should solve a
bunch of other issues where gnome-terminal is expected when GNOME is the DE.

[1]: https://gitlab.gnome.org/GNOME/glib/-/issues/338
[2]: https://gitlab.gnome.org/GNOME/glib/-/blob/master/gio/gdesktopappinfo.c#L2581