Fix scripts launched from incorrect directory
Until recently Desktop Entry files were executing the Exec line from the
directory where the opened file is located, but apparently this changed
recently and the Exec line is now always executed from the $HOME directory.

This commit fixes that by first cd'ing into the script's directory before
exec'ing it.

The new Exec line should also make it compliant with the Desktop File
specification, where it is mentioned that field codes must not be quoted.