Fix statusline color not correct sometimes
In some situations, like opening a file from NERDTree, the statusline color was
not properly updated, and instead it kept the color of the previous pane.

It was due to an incorrect assumption about what the "current value" is when
opening a new pane: it was assumed that it was the default value from the
theme, but instead it was the value from the previous pane.