Enable truecolors and improve graceful degradation
It seems that most recent terminal emulators support 24bit colors nowadays,
even if they don't advertise it via the TERM env var. The check is not perfect,
but if it fails we should also degrade better for terminals that support 256
colors (relying a bit less on having configured a Solarized theme), but using
the default theme as fallback for terminals that support even fewer colors.

The vanilla wombat theme has been removed as it requires truecolors anyways,
and is way less comprehensive than wombat256.