Streamline draw_tab logic
The logic that came from kitty's draw_tab_with_powerline was too convoluted,
with parts I didn't manage to grasp. It should now be pretty easy to follow.

This change improves the drawing of the first and last tabs (the title gets
drawn closer to the edge), and fixes the last tab being too wide (due to the
space reserved for a separator that doesn't get drawn).