Fix a bunch of bugs and edge cases
Now I'm hitting hard limitations, I can't just work around kitty's weirdness.
Sadly I probably won't be able to make broadcasting as good as Terminator's
with kitty's current API.

Not being able to auto-update the targets when windows are created is a big
sad. It would also have been nice to be able to style target windows' border,
but inactive border color is a global and can't be changed for individual
windows. Visually selecting targets when launching the broadcast would be nice,
but the `select-window` kitten is kind of broken for the use-case of selecting
multiple windows. I could add a way to toggle individual windows after the
broadcast has started, but I don't think I have it in me to do it :'(